requirements.txt
requests
beautifulsoup4
ultralytics
opencv-python
label-studio
pyyaml
